+++ packages/datagrid-vue-app/src/stage/dataGridTableStageBody.types.ts
@@ -129,6 +129,7 @@ export interface DataGridTableStagePinnedPaneRenderApi extends DataGridTableStag
   paneRowStyle: (row: DataGridTableStageBodyRow, rowOffset: number, paneWidth: number) => CSSProperties
   isFullRowSelectionSafe: (rowOffset: number) => boolean
   rowIndexColumnStyle: CSSProperties
+  rowIndexCellClasses: (row: DataGridTableStageBodyRow, rowOffset: number) => Record<string, boolean>
   rowIndexCellStyle: (row: DataGridTableStageBodyRow, rowOffset: number) => CSSProperties
   rowIndexTabIndex: (row: DataGridTableStageBodyRow) => number
   handleRowIndexClickSafe: (row: DataGridTableStageBodyRow, rowOffset: number, event: MouseEvent) => void